move mocking madness from v1_0 into its own namespace.
The first rule of unit testing is you DO NOT test outside of your unit
The second rule of unit testing is *YOU* *DO* *NOT* test outside of your unit

Given that :
 - decorators are complicated
 - mocking is hard to get right
 - overwriting namespaces is never a good idea

...mocking decorators by forcefully overwriting their namespace can only
lead to MADNESS. (It also leads to memory leaks. THE PAIIIIIIIIIN)

The decorator mocks in v1_0 were causing very bad side effects in the tests from
v1_1. Tests that were supposed to pass were failing. Tests that failed
were never seen.

After passing a whole morning dancing with the devil and slinging code
like an abstract painter who throws paint on a wall and calls it
'art', I have concluded that the simplest way out of this mayhem is to
simply copy the imports into seperate files, thus duplicating the
overwritten namespaces and moving them out of harms way
(remember the zen of python : "Namespaces are one honking great idea --
 let's do more of those!")