don't auto generate caller ids on user
Trying to magically update the caller id is a bad idea.
Client should update caller id when he needs to.
This means that when updating the firstname or lastname
of a user, a client should also update the caller id.
The only time a caller id will be auto generated is when
creating a user only if he doesn't already have one.