workaround for double identity map conflict
sqlalchemy raises an error because the sip object and the
sip relationship on the line are both exactly the same database
row, but they have been dissociated from the session.
These objects have been dissociated to avoid bugs by unwanted session
flushes in the dao.

This problem will be fixed when transaction management
has been fixed in the DAO (hopefully by next sprint)