integration tests: fix line-device association
When a device has at least 1 line associated to it, there is some
expectation that the config ID associated to the device is the same as
the device ID. But this is not handled correctly in xivo-confd, since
associating a line to a device in XiVO is an operation that only the
web-interface does, i.e. this is not completely implemented in
xivo-confd. So add a method that fix the expectation when associating a
line to a device in the integration tests.