user_import: use incall_extension service
The association of incall with line_extension has been removed. So the
incall_extension need to be used.
Also change the concept of incall in user_import. Before an incall was
an extension. Now an incall is a resource on which we can associate an
extension.

An little helper has been added in the integration tests to put an user
as destination of an incall.