parking_lot: add default value to timeout
reason: we cannot put the default value in dao (server_default=45),
since the column accept None value.
When create ParkingLot (sqlalchemy) object, there are no distinction
between:
ParkingLot(timeout=None)
ParkingLot()
The final value of timeout will always be 45.
In sqlalchemy 1.1 there are evaluates_none() to fix this problem. But
for the moment, it's simpler to set the default here.