line_device: remove usage of line_service
reason: When associate/dissociate line_device, we do not want a message
on the bus that is sent by updating line. We just want a line_device
message.

But the real reason behind this, it's that in the next commit, the
line_device service will be imported in line_service. So the circular
dependency need to be removed.