sound: return file without format/language first
reason: We cannot set format/language to None in the query string when
getting file.
ex:
GIVEN:
  * /path/category/toto.wav
  * /path/category/fr_FR/toto.wav

We want to get the first sound toto.wav:

WHEN GET /sounds/category/files/toto

But since we cannot specify the language to None, the result can be the
second toto.wav.

So we need to sort results before sending it.