refactor: avoid using the same logic when reading and creating resources
When creating a resource the model is checked to see if it has a tenant_uuid field.
When reading a resource the has_tenant_uuid field can be set on the resource to indicate
that the tenant has a tenant_uuid field. This is useful for resources with a tenant_uuid
that is "computed" from another object.

For example:

the tenant_uuid of a voicemail is the tenant_uuid of its context. The Voicemail object
does not accept a tenant_uuid on its constructor.