tests: fix python3.5 bug with csv.DictWriter
See https://bugs.python.org/issue32255
Since we always use csv.DictWriter with more than one column in the
production code (confd and dird), it is not really important. All
csv.DictWriter are used to expose data with many columns. But for the
tests, we use a custom DictWriter that can take one column. So to fix
it, we can simply add an optional column to avoid writing csv with one
column.