devices: add the device id on the corrupt device error
When you have a device with a config that does not exists anymore you GET a 500
when doing a GET on /devices with no indication of the bad device. From that
point you need to have a good understanding of wazo-provd-cli to be able to find
that corrupted device and remove it.