Remove Flask app context from syncdb
When the tenant deletion is called from the CLI (syncdb) the Flask app is not running.
So the sysconfd client must be initialized without using a Flask app.

But the Flask app context is used in the method called when the auth_tenant_deleted event
is consumed, otherwise a 'RuntimeError: Working outside of application context' is raised.