validate position from park position func key
note: we won't add more validation (ex: if parking lot is associated or
not to an extension). Because even if an admin change the slots range,
this function key can be wrong and we don't want to delete the function
key in a such situation.
This validation is intended to help user, not to constraint his workflow